Update shmatrix to deal with screen/shells.
Previously if you already had an env active and opened a screen you'd
end up with an active virtualenv and no way to deactivate it (because of
the way virtualenv works). This now checks if you're in an env but have
no deactivation command and automatically re-sources the activate
command. As a side effect, because we now need to check if the directory
you're in at load-time is a venv, if your homedir is a virtualenv it'll
work. Obscure.